refactor(reborn-memory): split lib.rs into focused modules
Phase 2 of #3118. Make the crate agent-friendly before the
deeper native-substrate changes by splitting the 3801-line
crates/ironclaw_memory/src/lib.rs into 13 focused modules. No
behavior change: every public name keeps its identity, the four
test files under crates/ironclaw_memory/tests/ are untouched, and
all 42 existing tests still pass.

New layout:

  src/lib.rs         (34 LoC)  re-export hub + crate docs
  src/path.rs        scope, path, parsing, validation, error helpers
  src/metadata.rs    DocumentMetadata, .config inheritance
  src/embedding.rs   EmbeddingProvider + vector helpers
  src/schema.rs      JSON schema validation
  src/search.rs      search request/result + RRF/weighted fusion
  src/chunking.rs    ChunkConfig + chunk_document + content_sha256
  src/backend.rs     MemoryBackend + RepositoryMemoryBackend
  src/filesystem.rs  RootFilesystem adapters
  src/indexer.rs     MemoryDocumentIndexer + ChunkingIndexer
  src/repo/mod.rs    MemoryDocumentRepository + shared repo helpers
  src/repo/in_memory.rs
  src/repo/libsql.rs        (cfg(libsql))
  src/repo/postgres.rs      (cfg(postgres))

Visibility rules:

- Public items keep their public names (verified against the
  imports in all four crates/ironclaw_memory/tests/*.rs files).
- Helpers shared across modules become pub(crate).
- libSQL- and Postgres-only schema constants and helpers stay
  inside their respective repo files.

Verification (all green):

  cargo fmt --all -- --check
  cargo clippy -p ironclaw_memory --all-targets --all-features -- -D warnings
  cargo test -p ironclaw_memory
  cargo test -p ironclaw_memory --features libsql
  cargo test -p ironclaw_memory --all-features
  cargo test -p ironclaw_architecture reborn_crate_dependency_boundaries_hold
  python3 scripts/check_no_panics.py --base origin/reborn-integration --head HEAD
  git diff --check origin/reborn-integration

This split is the foundation for the follow-up PRs that add the
native reborn_memory_* schema, libSQL/Postgres repositories on the
native schema, ported semantic/search/versioning tests, and host
wiring (#3118 phases 3-7).
