+++ crates/ironclaw_memory/src/filesystem.rs
@@ -1,0 +1,292 @@
+//! Memory-document `RootFilesystem` adapters.
+
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use ironclaw_filesystem::{
+    DirEntry, FileStat, FileType, FilesystemError, FilesystemOperation, RootFilesystem,
+};
+use ironclaw_host_api::VirtualPath;
+
+use crate::backend::{MemoryBackend, MemoryContext};
+use crate::indexer::MemoryDocumentIndexer;
+use crate::path::{
+    MemoryDocumentPath, MemoryDocumentScope, ParsedMemoryPath, memory_error, memory_not_found,
+};
+use crate::repo::{MemoryDocumentRepository, memory_direct_children};
+
+/// [`RootFilesystem`] adapter exposing any [`MemoryBackend`] as `/memory` files.
+pub struct MemoryBackendFilesystemAdapter {
+    backend: Arc<dyn MemoryBackend>,
+}
+
+impl MemoryBackendFilesystemAdapter {
+    pub fn new<B>(backend: Arc<B>) -> Self
+    where
+        B: MemoryBackend + 'static,
+    {
+        let backend: Arc<dyn MemoryBackend> = backend;
+        Self { backend }
+    }
+
+    pub fn from_dyn(backend: Arc<dyn MemoryBackend>) -> Self {
+        Self { backend }
+    }
+
+    fn ensure_file_documents(
+        &self,
+        path: &VirtualPath,
+        operation: FilesystemOperation,
+    ) -> Result<(), FilesystemError> {
+        if self.backend.capabilities().file_documents {
+            Ok(())
+        } else {
+            Err(memory_error(
+                path.clone(),
+                operation,
+                "memory backend does not support file documents",
+            ))
+        }
+    }
+
+    fn parse_file_path(
+        &self,
+        path: &VirtualPath,
+        operation: FilesystemOperation,
+    ) -> Result<MemoryDocumentPath, FilesystemError> {
+        let parsed = ParsedMemoryPath::from_virtual_path(path, operation)?;
+        let Some(relative_path) = parsed.relative_path else {
+            return Err(memory_error(
+                path.clone(),
+                operation,
+                "memory document path must include a file path after project id",
+            ));
+        };
+        Ok(MemoryDocumentPath {
+            scope: parsed.scope,
+            relative_path,
+        })
+    }
+}
+
+#[async_trait]
+impl RootFilesystem for MemoryBackendFilesystemAdapter {
+    async fn read_file(&self, path: &VirtualPath) -> Result<Vec<u8>, FilesystemError> {
+        self.ensure_file_documents(path, FilesystemOperation::ReadFile)?;
+        let document_path = self.parse_file_path(path, FilesystemOperation::ReadFile)?;
+        let context = MemoryContext::new(document_path.scope().clone());
+        self.backend
+            .read_document(&context, &document_path)
+            .await?
+            .ok_or_else(|| memory_not_found(path.clone(), FilesystemOperation::ReadFile))
+    }
+
+    async fn write_file(&self, path: &VirtualPath, bytes: &[u8]) -> Result<(), FilesystemError> {
+        self.ensure_file_documents(path, FilesystemOperation::WriteFile)?;
+        let document_path = self.parse_file_path(path, FilesystemOperation::WriteFile)?;
+        let context = MemoryContext::new(document_path.scope().clone());
+        self.backend
+            .write_document(&context, &document_path, bytes)
+            .await
+    }
+
+    async fn list_dir(&self, path: &VirtualPath) -> Result<Vec<DirEntry>, FilesystemError> {
+        self.ensure_file_documents(path, FilesystemOperation::ListDir)?;
+        let parsed = ParsedMemoryPath::from_virtual_path(path, FilesystemOperation::ListDir)?;
+        let context = MemoryContext::new(parsed.scope.clone());
+        let documents = self.backend.list_documents(&context, &parsed.scope).await?;
+        if let Some(relative_path) = parsed.relative_path.as_deref()
+            && documents
+                .iter()
+                .any(|document| document.relative_path() == relative_path)
+        {
+            return Err(memory_error(
+                path.clone(),
+                FilesystemOperation::ListDir,
+                "not a directory",
+            ));
+        }
+        memory_direct_children(path, parsed.relative_path.as_deref(), documents)
+    }
+
+    async fn stat(&self, path: &VirtualPath) -> Result<FileStat, FilesystemError> {
+        self.ensure_file_documents(path, FilesystemOperation::Stat)?;
+        let parsed = ParsedMemoryPath::from_virtual_path(path, FilesystemOperation::Stat)?;
+        let context = MemoryContext::new(parsed.scope.clone());
+        let documents = self.backend.list_documents(&context, &parsed.scope).await?;
+        if let Some(relative_path) = parsed.relative_path.as_deref() {
+            if let Some(document) = documents
+                .iter()
+                .find(|document| document.relative_path() == relative_path)
+            {
+                let len = self
+                    .backend
+                    .read_document(&context, document)
+                    .await?
+                    .map(|bytes| bytes.len() as u64)
+                    .unwrap_or(0);
+                return Ok(FileStat {
+                    path: path.clone(),
+                    file_type: FileType::File,
+                    len,
+                });
+            }
+            let directory_prefix = format!("{relative_path}/");
+            if documents
+                .iter()
+                .any(|document| document.relative_path().starts_with(&directory_prefix))
+            {
+                return Ok(FileStat {
+                    path: path.clone(),
+                    file_type: FileType::Directory,
+                    len: 0,
+                });
+            }
+            return Err(memory_not_found(path.clone(), FilesystemOperation::Stat));
+        }
+
+        if documents.is_empty() {
+            return Err(memory_not_found(path.clone(), FilesystemOperation::Stat));
+        }
+        Ok(FileStat {
+            path: path.clone(),
+            file_type: FileType::Directory,
+            len: 0,
+        })
+    }
+}
+
+/// [`RootFilesystem`] backend exposing DB-backed memory documents as virtual files.
+pub struct MemoryDocumentFilesystem {
+    repository: Arc<dyn MemoryDocumentRepository>,
+    indexer: Option<Arc<dyn MemoryDocumentIndexer>>,
+}
+
+impl MemoryDocumentFilesystem {
+    pub fn new<R>(repository: Arc<R>) -> Self
+    where
+        R: MemoryDocumentRepository + 'static,
+    {
+        Self {
+            repository,
+            indexer: None,
+        }
+    }
+
+    pub fn with_indexer<I>(mut self, indexer: Arc<I>) -> Self
+    where
+        I: MemoryDocumentIndexer + 'static,
+    {
+        self.indexer = Some(indexer);
+        self
+    }
+
+    fn parse_file_path(
+        &self,
+        path: &VirtualPath,
+        operation: FilesystemOperation,
+    ) -> Result<MemoryDocumentPath, FilesystemError> {
+        let parsed = ParsedMemoryPath::from_virtual_path(path, operation)?;
+        let Some(relative_path) = parsed.relative_path else {
+            return Err(memory_error(
+                path.clone(),
+                operation,
+                "memory document path must include a file path after project id",
+            ));
+        };
+        Ok(MemoryDocumentPath {
+            scope: parsed.scope,
+            relative_path,
+        })
+    }
+
+    async fn list_for_scope(
+        &self,
+        scope: &MemoryDocumentScope,
+    ) -> Result<Vec<MemoryDocumentPath>, FilesystemError> {
+        self.repository.list_documents(scope).await
+    }
+}
+
+#[async_trait]
+impl RootFilesystem for MemoryDocumentFilesystem {
+    async fn read_file(&self, path: &VirtualPath) -> Result<Vec<u8>, FilesystemError> {
+        let document_path = self.parse_file_path(path, FilesystemOperation::ReadFile)?;
+        self.repository
+            .read_document(&document_path)
+            .await?
+            .ok_or_else(|| memory_not_found(path.clone(), FilesystemOperation::ReadFile))
+    }
+
+    async fn write_file(&self, path: &VirtualPath, bytes: &[u8]) -> Result<(), FilesystemError> {
+        let document_path = self.parse_file_path(path, FilesystemOperation::WriteFile)?;
+        self.repository
+            .write_document(&document_path, bytes)
+            .await?;
+        if let Some(indexer) = &self.indexer {
+            let _ = indexer.reindex_document(&document_path).await;
+        }
+        Ok(())
+    }
+
+    async fn list_dir(&self, path: &VirtualPath) -> Result<Vec<DirEntry>, FilesystemError> {
+        let parsed = ParsedMemoryPath::from_virtual_path(path, FilesystemOperation::ListDir)?;
+        let documents = self.list_for_scope(&parsed.scope).await?;
+        if let Some(relative_path) = parsed.relative_path.as_deref()
+            && documents
+                .iter()
+                .any(|document| document.relative_path() == relative_path)
+        {
+            return Err(memory_error(
+                path.clone(),
+                FilesystemOperation::ListDir,
+                "not a directory",
+            ));
+        }
+        memory_direct_children(path, parsed.relative_path.as_deref(), documents)
+    }
+
+    async fn stat(&self, path: &VirtualPath) -> Result<FileStat, FilesystemError> {
+        let parsed = ParsedMemoryPath::from_virtual_path(path, FilesystemOperation::Stat)?;
+        let documents = self.list_for_scope(&parsed.scope).await?;
+        if let Some(relative_path) = parsed.relative_path.as_deref() {
+            if let Some(document) = documents
+                .iter()
+                .find(|document| document.relative_path() == relative_path)
+            {
+                let len = self
+                    .repository
+                    .read_document(document)
+                    .await?
+                    .map(|bytes| bytes.len() as u64)
+                    .unwrap_or(0);
+                return Ok(FileStat {
+                    path: path.clone(),
+                    file_type: FileType::File,
+                    len,
+                });
+            }
+            let directory_prefix = format!("{relative_path}/");
+            if documents
+                .iter()
+                .any(|document| document.relative_path().starts_with(&directory_prefix))
+            {
+                return Ok(FileStat {
+                    path: path.clone(),
+                    file_type: FileType::Directory,
+                    len: 0,
+                });
+            }
+            return Err(memory_not_found(path.clone(), FilesystemOperation::Stat));
+        }
+
+        if documents.is_empty() {
+            return Err(memory_not_found(path.clone(), FilesystemOperation::Stat));
+        }
+        Ok(FileStat {
+            path: path.clone(),
+            file_type: FileType::Directory,
+            len: 0,
+        })
+    }
+}
